增加了解除口球抽奖

+++ src/main/java/com/piggy/quincy/component/MiraiApiHttpComponent.java
@@ -269,7 +269,7 @@ public class MiraiApiHttpComponent {
      * @return Response
      * @throws IOException
      */
-    public Response recall(String sessionKey, Long target) throws IOException {
+    public Response recall(String sessionKey, Integer target) throws IOException {
         JSONObject jsonObject = new JSONObject();
         jsonObject.put("sessionKey", sessionKey);
         jsonObject.put("target", target);
